requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
